+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to this project will be documented in this file.
 - Contact message creation on oxe-web-community by correcting database structure via migration
 - Allow modification of user email via oxe-web-admin
 - SQL querying requiring GROUP BY with the dockerized madiadb
+- Show all optional field for article dialog, whatever the article type on oxe-web-admin
 
 ## [1.15.5] - 2023-04-04