+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.16.0] - XXXX-XX-XX
+## [1.16.0] - 2023-07-11
 
 ### Changed